:art: Componentize library and sharing interactions
