+++ src/app/cube/details/details.component.scss
@@ -72,206 +72,7 @@ $content-width: calc(#{$max-width} - #{$sidebar-width});
       margin-top: -100px;
       z-index: 2;
 
-      .inner {
-        background: white;
-        border-radius: 2px;
-        box-shadow: 0 0 12px 1px rgba(0, 0, 0, 0.03);
-        padding: 20px;
-        margin-bottom: 20px;
-
-        .download-warning {
-          color: white;
-          background: linear-gradient(45deg, #fea200, #ffd000);
-          padding: 8px 12px;
-          border-radius: 2px 2px 0 0;
-          font-size: 18px;
-          text-align: center;
-          margin-bottom: 10px;
-          text-shadow: 0 0 2px rgba(0, 0, 0, 0.2);
-          position: absolute;
-          top: 0;
-          left: 0;
-          right: 0;
-
-          .svg-inline--fa {
-            margin-right: 10px;
-          }
-        }
-
-        .btn-group.no-download {
-          margin-top: 40px;
-        }
-
-        .button {
-          overflow: hidden;
-          text-align: center;
-          font-size: 16px;
-          margin-bottom: 10px;
-          position: relative;
-
-          &.disabled {
-            cursor: default !important;
-            opacity: 0.4 !important;
-            border-color: $dark-text !important;
-            pointer-events: none;
-            font-size: 14px;
-
-            &.good {
-              background: $dark-text !important;
-            }
-
-            &.neutral {
-              color: $dark-text !important;
-            }
-          }
-
-          &.saved {
-            cursor: default !important;
-            pointer-events: none;
-            border-color: transparent;
-            color: $green;
-            box-shadow: none;
-          }
-        }
-
-        .login-msg {
-          margin-top: 20px;
-          font-size: 16px;
-          color: $dark-text;
-
-          span {
-            color: $blue-accent;
-            text-decoration: underline;
-            cursor: pointer;
-          }
-        }
-      }
-
-      .details-metrics {
-        display: flex;
-        flex-direction: row;
-        justify-content: center;
-        color: lighten($dark-text, 20);
-        margin-top: 30px;
-
-        & > * {
-
-          span {
-            position: relative;
-
-            &.animate {
-              animation: newSaveOrDownload 0.8s ease;
-            }
-          }
-
-          &:first-child {
-            margin-right: 20px;
-          }
-        }
-
-        .svg-inline--fa {
-          margin-right: 5px;
-        }
-      }
-
-      .details-share {
-        margin-top: 30px;
-
-        .share-title {
-          position: relative;
-
-          &:before {
-            content: 'Share';
-            color: $dark-text;
-            text-align: center;
-            padding: 0 10px;
-            background: white;
-            position: absolute;
-            z-index: 2;
-            left: 30%;
-            right: 30%;
-            bottom: -7px;
-          }
-
-          &:after {
-            content: '';
-            position: absolute;
-            height: 1px;
-            background: lighten($dark-text, 40);
-            width: 100%;
-            left: 0;
-            bottom: 0;
-          }
-        }
-
-        .share-icons {
-          padding-top: 20px;
-          display: flex;
-          flex-direction: row;
-          justify-content: space-between;
-
-          button {
-            border: 0;
-            -webkit-appearance: none;
-            -moz-appearance: none;
-            outline: 0;
-          }
-
-          a {
-            color: white !important;
-            text-decoration: none !important;
-          }
-
-          & > button {
-            min-width: 30px;
-            height: 30px;
-            background: $blue-accent;
-            border-radius: 3px;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            color: white !important;
-            font-size: 18px;
-            transition: all 0.2s ease;
-            transform: scale(1);
-            cursor: pointer;
-            flex-shrink: 0;
-            flex-grow: 0;
-
-            &:hover:not(.animated) {
-              transform: scale(1.1);
-            }
-
-            &:active:hover {
-              transform: scale(0.95);
-            }
-
-            &.animated {
-              animation: jiggle 0.4s linear;
-            }
-
-            &.facebook {
-              background: #3b5998;
-            }
-
-            &.twitter {
-              background: #1da1f2;
-            }
-
-            &.linkedin {
-              background: #0077b5;
-            }
-
-            &.email {
-              background: $error-red;
-            }
-
-            &.copylink {
-              background: $green;
-            }
-          }
-        }
-      }
+      
     }
   }
 }
@@ -311,19 +112,6 @@ $content-width: calc(#{$max-width} - #{$sidebar-width});
     .button {
       width: 100% !important;
     }
-  }
-}
-
-.author {
-  .link {
-    color: #fff;
-    opacity: 1;
-    font-weight: 400;
-  }
-
-  .link:hover {
-    cursor: pointer;
-    text-decoration: underline;
   }
 }
